+++ playlet-web/src/lib/VideoFeed/PlaylistThumbnail.svelte
@@ -32,7 +32,7 @@
       }
     }
     if (url === "") {
-      url = `${invidiousInstance}/vi/-----------/mqdefault.jpg`;
+      url = `https://i.ytimg.com/vi/-----------/mqdefault.jpg`;
     }
     thumbnailUrl = url;
   }